Integrate Python union regressions into existing input tests
