--- tests/test_input_model_union.py
@@ -1,132 +1,0 @@
-"""CLI/API regressions for source annotations attached to union schema choices."""
-
-from __future__ import annotations
-
-import json
-from pathlib import Path
-
-import pytest
-from pydantic import ValidationError
-
-from datamodel_code_generator import DataModelType, GenerateConfig, InputFileType, generate
-from datamodel_code_generator.format import Formatter
-from datamodel_code_generator.input_model import load_model_schema
-from tests.conftest import assert_inputs_not_mutated, assert_output
-from tests.data.python.input_model import union_annotations
-from tests.data.python.input_model.union_runtime import CASES, VALUES, describe
-from tests.main.conftest import _generated_model, run_main_with_args
-
-EXPECTED = Path(__file__).parent / "data" / "expected" / "main" / "input_model" / "union_annotations"
-INPUT_MODULE = "tests.data.python.input_model.union_annotations"
-
-
-@pytest.mark.parametrize("model_name", CASES)
-@pytest.mark.parametrize("entrypoint", ["cli", "api"])
-@pytest.mark.parametrize("formatter", ["external", "builtin"])
-def test_input_model_union_branch_provenance(model_name: str, entrypoint: str, formatter: str, tmp_path: Path) -> None:
-    """Keep union order, exact generated bytes, and source runtime acceptance."""
-    source = f"{INPUT_MODULE}:{model_name}"
-    output = tmp_path / "output.py"
-    formatters = ["builtin"] if formatter == "builtin" else ["black", "isort"]
-    base_class = f"{INPUT_MODULE}.ArbitraryBase" if model_name == "MultipleUnsupported" else None
-    extra_args = ["--base-class", base_class] if base_class else []
-    if entrypoint == "cli":
-        run_main_with_args(
-            [
-                "--input-model",
-                source,
-                "--output",
-                str(output),
-                "--disable-timestamp",
-                "--formatters",
-                *formatters,
-                *extra_args,
-            ],
-            use_parsed_source_cache=False,
-        )
-    else:
-        generate(
-            load_model_schema([source], InputFileType.JsonSchema),
-            config=GenerateConfig(
-                base_class=base_class or "",
-                input_file_type=InputFileType.JsonSchema,
-                input_filename="<stdin>",
-                output_model_type=DataModelType.PydanticV2BaseModel,
-                disable_timestamp=True,
-                output=output,
-                formatters=[Formatter(value) for value in formatters],
-            ),
-        )
-    expected_name = f"{model_name}_builtin" if formatter == "builtin" and base_class else model_name
-    assert_output(output.read_text(), EXPECTED / f"{expected_name}.py")
-    schema = load_model_schema([source], InputFileType.JsonSchema)
-    assert_output(json.dumps(schema, indent=2), EXPECTED / f"{model_name}_schema.txt")
-    records = {}
-    with _generated_model(output, "_generated_union_annotations", model_name) as generated:
-        for label in CASES[model_name]:
-            payload = {"value": VALUES[label]}
-            observations = {}
-            for name, model in (("source", getattr(union_annotations, model_name)), ("generated", generated)):
-                with assert_inputs_not_mutated(payload):
-                    try:
-                        result = model.model_validate(payload)
-                    except ValidationError:
-                        observations[name] = "rejected"
-                    else:
-                        observations[name] = describe(result.model_dump())
-            records[label] = observations
-    assert_output(json.dumps(records, indent=2), EXPECTED / f"{model_name}_runtime.txt")
-
-
-@pytest.mark.parametrize("model_name", ["Serializable", "Mixed", "Container", "Nested"])
-@pytest.mark.parametrize("entrypoint", ["cli", "api"])
-@pytest.mark.parametrize("formatter", ["external", "builtin"])
-def test_input_model_union_caller_extensions(model_name: str, entrypoint: str, formatter: str, tmp_path: Path) -> None:
-    """Consume generated provenance without discarding similarly named user metadata."""
-    from tests.data.python.input_model import union_extensions
-
-    source = f"tests.data.python.input_model.union_extensions:{model_name}"
-    output = tmp_path / "output.py"
-    formatters = ["builtin"] if formatter == "builtin" else ["black", "isort"]
-    if entrypoint == "cli":
-        run_main_with_args([
-            "--input-model",
-            source,
-            "--output",
-            str(output),
-            "--disable-timestamp",
-            "--formatters",
-            *formatters,
-        ])
-    else:
-        generate(
-            load_model_schema([source], InputFileType.JsonSchema),
-            config=GenerateConfig(
-                input_file_type=InputFileType.JsonSchema,
-                input_filename="<stdin>",
-                disable_timestamp=True,
-                output=output,
-                formatters=[Formatter(value) for value in formatters],
-            ),
-        )
-    assert_output(output.read_text(), EXPECTED / f"extension_{model_name}.py")
-    schema = load_model_schema([source], InputFileType.JsonSchema)
-    assert_output(json.dumps(schema, indent=2), EXPECTED / f"extension_{model_name}_schema.txt")
-    if model_name == "Serializable":
-        assert_output(
-            json.dumps(union_extensions.Serializable.model_json_schema(), indent=2),
-            EXPECTED / f"extension_{model_name}_schema.txt",
-        )
-    with _generated_model(output, "_generated_union_extensions", model_name) as generated:
-        for model in (getattr(union_extensions, model_name), generated):
-            observations = []
-            for value in (1, "text", int, None, [int]):
-                payload = {"value": value}
-                with assert_inputs_not_mutated(payload):
-                    try:
-                        result = model.model_validate(payload)
-                    except ValidationError:
-                        observations.append("rejected")
-                    else:
-                        observations.append(describe(result.model_dump()))
-            assert_output(json.dumps(observations, indent=2), EXPECTED / f"extension_{model_name}_runtime.txt")
